FWA-7 - EDIT and DELETE button does not work in users page

+++ src/main/java/com/furidaweb/server/service/user/impl/UserServiceImpl.java
@@ -5,9 +5,11 @@ import com.furidaweb.server.repository.UserRepository;
 import com.furidaweb.server.service.user.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @RequiredArgsConstructor
 @Service
@@ -32,7 +34,14 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUserById(int id) {
+    public void deleteUserById(String username, int id) {
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
+
+        if (Objects.equals(user.getUsername(), username)) {
+            throw new IllegalArgumentException("Logged in user cannot be deleted.");
+        }
+
         userRepository.deleteById(id);
     }
